--- a/CSharp/Clipper2Lib.Benchmark/Program.cs
+++ b/CSharp/Clipper2Lib.Benchmark/Program.cs
@@ -1,0 +1,12 @@
+using BenchmarkDotNet.Running;
+
+namespace Clipper2Lib.Benchmark
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            var summary = BenchmarkRunner.Run<Benchmarks>();
+        }
+    }
+}