--- a/CSharp/Clipper2Lib.sln
+++ b/CSharp/Clipper2Lib.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32014.148
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Clipper2Lib", "Clipper2Lib\Clipper2Lib.csproj", "{4CE00B81-F1B2-4012-9BBD-B0DB6BF8CAD3}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Clipper2Lib.Benchmark", "Clipper2Lib.Benchmark\Clipper2Lib.Benchmark.csproj", "{F8A168C6-286A-4ECF-8750-C505ADBCB7D5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Clipper2LibExample", "Clipper2LibExample\Clipper2LibExample.csproj", "{4BA8DEFE-0678-4C41-9B16-3A4639DFB838}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4CE00B81-F1B2-4012-9BBD-B0DB6BF8CAD3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4CE00B81-F1B2-4012-9BBD-B0DB6BF8CAD3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4CE00B81-F1B2-4012-9BBD-B0DB6BF8CAD3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4CE00B81-F1B2-4012-9BBD-B0DB6BF8CAD3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F8A168C6-286A-4ECF-8750-C505ADBCB7D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F8A168C6-286A-4ECF-8750-C505ADBCB7D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F8A168C6-286A-4ECF-8750-C505ADBCB7D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F8A168C6-286A-4ECF-8750-C505ADBCB7D5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4BA8DEFE-0678-4C41-9B16-3A4639DFB838}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4BA8DEFE-0678-4C41-9B16-3A4639DFB838}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4BA8DEFE-0678-4C41-9B16-3A4639DFB838}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4BA8DEFE-0678-4C41-9B16-3A4639DFB838}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EB14E46A-0760-411E-B991-4E0B4DBBFF93}
+	EndGlobalSection
+EndGlobal