--- a/CSharp/Clipper2Lib/Clipper2Lib.csproj
+++ b/CSharp/Clipper2Lib/Clipper2Lib.csproj
@@ -1,0 +1,8 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.1</TargetFramework>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+</Project>