--- a/CSharp/Clipper2LibExample/Clipper2LibExample.csproj
+++ b/CSharp/Clipper2LibExample/Clipper2LibExample.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Clipper2Lib\Clipper2Lib.csproj" />
+  </ItemGroup>
+
+</Project>