… CSharp/.gitignore
… CSharp/Clipper2Lib/Clipper.Core.cs
… CSharp/Clipper2Lib/Clipper.cs