### Goal
Increase performance of the Csharp builds with minimal impact to original algorithm design.

###
- [x] Inline methods for performance.
- [ ] Utilize dotnet methodology for [readonly struct](https://docs.microsoft.com/en-us/dotnet/csharp/write-safe-efficient-code)
- [ ] Review options for internally reusing Lists or direct array assignment.

### Results.
``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19044
Intel Core i7-10700 CPU 2.90GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=6.0.201
  [Host]     : .NET Core 5.0.15 (CoreCLR 5.0.1522.11506, CoreFX 5.0.1522.11506), X64 RyuJIT
  DefaultJob : .NET Core 5.0.15 (CoreCLR 5.0.1522.11506, CoreFX 5.0.1522.11506), X64 RyuJIT
```

<details>
  <summary>Baseline 96f3cc5be10c5e194b1fbce530ea17252e214451</summary>

|                    Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|----------:|----------:|-------:|-------:|------:|----------:|
|   Union_Null_Clip_EvenOdd | 1.949 μs | 0.0569 μs | 0.0031 μs | 0.4997 | 0.0038 |     - |    4.1 KB |
|   Union_Null_Clip_NonZero | 1.622 μs | 0.1272 μs | 0.0070 μs | 0.4234 | 0.0019 |     - |   3.46 KB |
|        Union_Clip_NonZero | 3.635 μs | 0.1214 μs | 0.0067 μs | 0.7362 | 0.0076 |     - |   6.02 KB |
|    Intersect_Clip_NonZero | 3.425 μs | 0.2235 μs | 0.0123 μs | 0.6981 | 0.0076 |     - |   5.71 KB |
|    Intersect_Clip_EvenOdd | 3.758 μs | 0.1699 μs | 0.0093 μs | 0.7668 |      - |     - |   6.27 KB |
|       InflatePaths_Delta5 | 1.691 μs | 0.1571 μs | 0.0086 μs | 0.5093 | 0.0038 |     - |   4.16 KB |
|  InflatePaths_DeltaMinus5 | 4.234 μs | 0.4420 μs | 0.0242 μs | 0.9995 | 0.0076 |     - |   8.18 KB |
|      InflatePaths_Delta10 | 5.100 μs | 0.1322 μs | 0.0072 μs | 1.1597 | 0.0153 |     - |   9.51 KB |
| InflatePaths_DeltaMinus10 | 6.292 μs | 0.1274 μs | 0.0070 μs | 1.3199 | 0.0229 |     - |   10.8 KB |
</details>
<details>
  <summary>Inlined methods a48fd27f3546720b43310f9e4c1b7667be6c4f2c</summary>

| **Diff** | Method                    | Mean                | Error         | Gen 0            | Gen 1             | Allocated        |
| -------- | ------------------------- | -------------------:| ------------- | ----------------:| -----------------:| ----------------:|
| Old      | Union_Null_Clip_EvenOdd   | 1.809 μs            | 0.0101 μs     | 0.4997           | 0.0038            | 4.1 KB           |
| **New**  |                           | **1.703 μs (-6%)**  | **0.0129 μs** | **0.5016 (0%)**  | **0.0057 (+50%)** | **4.1 KB (0%)**  |
| Old      | Union_Null_Clip_NonZero   | 1.697 μs            | 0.0227 μs     | 0.4234           | 0.0019            | 3.46 KB          |
| **New**  |                           | **1.558 μs (-8%)**  | **0.0190 μs** | **0.4234 (0%)**  | **0.0019 (0%)**   | **3.46 KB (0%)** |
| Old      | Union_Clip_NonZero        | 3.665 μs            | 0.0068 μs     | 0.7362           | 0.0076            | 6.02 KB          |
| **New**  |                           | **3.347 μs (-9%)**  | **0.0258 μs** | **0.7362 (0%)**  | **0.0076 (0%)**   | **6.02 KB (0%)** |
| Old      | Intersect_Clip_NonZero    | 3.457 μs            | 0.0166 μs     | 0.6981           | 0.0076            | 5.71 KB          |
| **New**  |                           | **3.205 μs (-7%)**  | **0.0245 μs** | **0.6981 (0%)**  | **0.0076 (0%)**   | **5.71 KB (0%)** |
| Old      | Intersect_Clip_EvenOdd    | 3.792 μs            | 0.0233 μs     | 0.7629           | 0.0000            | 6.27 KB          |
| **New**  |                           | **3.494 μs (-8%)**  | **0.0244 μs** | **0.7668 (+1%)** | **0.0000**        | **6.27 KB (0%)** |
| Old      | InflatePaths_Delta5       | 1.713 μs            | 0.0126 μs     | 0.5093           | 0.0038            | 4.16 KB          |
| **New**  |                           | **1.593 μs (-7%)**  | **0.0064 μs** | **0.5093 (0%)**  | **0.0038 (0%)**   | **4.16 KB (0%)** |
| Old      | InflatePaths_DeltaMinus5  | 4.241 μs            | 0.0236 μs     | 0.9995           | 0.0076            | 8.18 KB          |
| **New**  |                           | **3.832 μs (-10%)** | **0.0312 μs** | **0.9995 (0%)**  | **0.0114 (+50%)** | **8.18 KB (0%)** |
| Old      | InflatePaths_Delta10      | 5.045 μs            | 0.0651 μs     | 1.1597           | 0.0153            | 9.51 KB          |
| **New**  |                           | **4.678 μs (-7%)**  | **0.0228 μs** | **1.1597 (0%)**  | **0.0153 (0%)**   | **9.51 KB (0%)** |
| Old      | InflatePaths_DeltaMinus10 | 6.517 μs            | 0.0747 μs     | 1.3199           | 0.0229            | 10.8 KB          |
| **New**  |                           | **5.858 μs (-10%)** | **0.0447 μs** | **1.3199 (0%)**  | **0.0229 (0%)**   | **10.8 KB (0%)** |

</details>