@philstopford 2022-04-04
Could you add some detail about the observed performance gain from this change? It would make the PR more immediately informative :)