@DJGosnell 2022-04-04
How opposed would you be to making the C# code consistent between the different clipper CS files?
`Clipper.Core.cs`, `Clipper.Offset.cs` and `Clipper.cs` uses two spaces while `Clipper.Engine.cs` uses mostly tabs with some spaces mixed in.  

Can I make these consistent between each of the source files?  Tabs or spaces don't matter much to me.  The dotnet community has mostly centered around spaces, but any good IDE worth its salt will work just fine either way.   The mixture just becomes irritating sometimes and this seems like a good time to address it.