@AngusJohnson 2022-04-05
> Could you add some detail about the observed performance gain from this change?

My admittedly brief testing has shown and performance improvement. 
Still, I don't want to discourage anyone trying 😁.