@philstopford 2022-04-05
> > Could you add some detail about the observed performance gain from this change?
> 
> My admittedly brief testing has shown any performance improvement. Still, I don't want to discourage anyone trying 😁.

That was really to DJGosnell - just to flesh out the PR, which he then did :)